added castle gates and  entrance mountain to level 3
Castle gates are required for level 3 as transitional measures between the village platforms and the final castle tile.

The mountains at the beginning is also essential where they are to bridge the narrative between level 2 (the cave) and level 3. As is will look like the players exit the cave via a mountain mine exit to the village.